add download image as png
